requirements.txt
# GCP
google-cloud-storage
# for training
python-dotenv
protobuf==3.20.0
mediapipe
numpy
accelerate==0.15.0
albumentations==1.3.0
altair==4.2.2
bitsandbytes==0.37.0
dadaptation==1.5
diffusers[torch]==0.10.2
easygui==0.98.3
einops==0.6.0
ftfy==6.1.1
lion-pytorch==0.0.6
opencv-python==4.7.0.68
pytorch-lightning==1.9.0
safetensors==0.2.6
tensorboard==2.10.1
tk==0.1.0
torch==1.13.1
torchvision==0.14.1
triton
transformers==4.26.0
# for BLIP captioning
fairscale==0.4.13
requests
timm==0.6.12
# tensorflow<2.11
huggingface-hub>=0.12.0
tensorflow==2.10.1
# For locon support
locon==0.0.2
xformers==0.0.16
#webui

blendmodes
basicsr
fonts
font-roboto
gfpgan
gradio==3.23
invisible-watermark
omegaconf
opencv-contrib-python
piexif
Pillow
realesrgan
scikit-image>=0.19
einops
jsonmerge
clean-fid
resize-right
torchdiffeq
kornia
lark
inflection
GitPython
torchsde
psutil
rich
